Put reset to default values functionality
